pulumi attempts in typescript and python
This will create a cluster, build the containers and deploy the
helm chart.

The typescript version is a bit more concise but I have been doing things
differently based on various online resources.

To use, download pulumi, cd to the repo and do `pulumi up`

If you're familiar with terraform, the "remote state" is handed out of
the box, on their website. That's nice. Also, I like the way that the
cluster declaration and the k8s declaration are integrated together in
a few lines of code. Also, it's pure python (or typescript) so the
flexibility is greated than terraform's dsl.

Cons:
the remote state is free for individual usage but will probably not be
free for a team.
It's not trivial to build your images and it uses the local docker. It
would be nice if it used kaniko (cloud-independent way of building your
containers directly on the target cluster), or amazon codebuild.

Next steps:

* decide on whether python or typescript
* add https endpoint for rpc auth and perhaps a managed redis instance
* add namespace support
* decide where to put the code. for now, I suggest a dedicated tq infra private repo
* add some github actions for infrastructure autodeploy