remove p2p service, inject listen address from pod
remove p2p service + add a bootstrap wait mechanism

The p2p protocol makes a bootstrap service unnecessary - peers
connect to each other directly, they don't need any frontend.

However there was a side effect of having a p2p service: the node
were connecting to the proxy which was establishing connections.

Removing the service cause peers to be stuck because they try to
connect directly to the bootstrap peer before it's ready.

To solve this, it's necessary to add a wait mechanism at first boot.

It delays launch of the other nodes until the bootstrap node is
accepting connections.