Restore ability to scale up number of nodes
To scale up, run mkchain again with a higher --number-of-nodes then run
`helm upgrade --install`

When running mkchain with a chain_name corresponding to helm values
files that already exist, it will preserve the timestamps and baking
accounts.

Scale up of nodes is permitted and works, scale up of number of
bakers is not permitted as it would not work.

Turning a regular chain into a zerotier chain works (I checked).